refactor: moved post validations to services

* refactor: moved post validations to services

+++ service/src/main/java/ar/edu/itba/paw/service/BuyOrderServiceImpl.java
@@ -1,9 +1,6 @@
 package ar.edu.itba.paw.service;
 
-import ar.edu.itba.paw.exceptions.ImageNotFoundException;
-import ar.edu.itba.paw.exceptions.NftNotFoundException;
-import ar.edu.itba.paw.exceptions.SellOrderNotFoundException;
-import ar.edu.itba.paw.exceptions.UserNotFoundException;
+import ar.edu.itba.paw.exceptions.*;
 import ar.edu.itba.paw.model.*;
 import ar.edu.itba.paw.persistence.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,7 +15,7 @@ import java.util.Optional;
 @Service
 public class BuyOrderServiceImpl implements BuyOrderService {
     private final BuyOrderDao buyOrderDao;
-    private final UserDao userDao;
+    private final UserService userService;
     private final SellOrderDao sellOrderDao;
     private final NftDao nftDao;
     private final ImageDao imageDao;
@@ -26,9 +23,9 @@ public class BuyOrderServiceImpl implements BuyOrderService {
     private final PurchaseService purchaseService;
 
     @Autowired
-    public BuyOrderServiceImpl(BuyOrderDao  buyOrderDao, UserDao userDao, SellOrderDao sellOrderDao, NftDao nftDao, ImageDao imageDao, MailingService mailingService, PurchaseService purchaseService) {
+    public BuyOrderServiceImpl(BuyOrderDao  buyOrderDao, UserService userService, SellOrderDao sellOrderDao, NftDao nftDao, ImageDao imageDao, MailingService mailingService, PurchaseService purchaseService) {
         this.buyOrderDao = buyOrderDao;
-        this.userDao = userDao;
+        this.userService = userService;
         this.sellOrderDao = sellOrderDao;
         this.nftDao = nftDao;
         this.imageDao = imageDao;
@@ -38,12 +35,15 @@ public class BuyOrderServiceImpl implements BuyOrderService {
 
     @Override
     public boolean create(int idSellOrder, BigDecimal price, int userId) {
+        if(userService.currentUserOwnsSellOrder(idSellOrder))
+            throw new UserNoPermissionException();
+
         if(buyOrderDao.create(idSellOrder, price, userId)){
             BuyOrder buyOrder = new BuyOrder(idSellOrder, price, userId);
             SellOrder sellOrder = sellOrderDao.getOrderById(idSellOrder).orElseThrow(SellOrderNotFoundException::new);
             Nft nft = nftDao.getNFTById(sellOrder.getNftId()).orElseThrow(NftNotFoundException::new);
-            User seller = userDao.getUserById(nft.getIdOwner()).orElseThrow(UserNotFoundException::new);
-            User bidder = userDao.getUserById(userId).orElseThrow(UserNotFoundException::new);
+            User seller = userService.getUserById(nft.getIdOwner()).orElseThrow(UserNotFoundException::new);
+            User bidder = userService.getUserById(userId).orElseThrow(UserNotFoundException::new);
             Image image = imageDao.getImage(nft.getIdImage()).orElseThrow(ImageNotFoundException::new);
             mailingService.sendOfferMail(bidder.getEmail(), seller.getEmail(), nft.getNftName(), nft.getId(), nft.getContractAddr(), buyOrder.getAmount(), image.getImage());
             return true;
@@ -56,7 +56,7 @@ public class BuyOrderServiceImpl implements BuyOrderService {
         List<BuyOrder> buyOrders = buyOrderDao.getOrdersBySellOrderId(offerPage, idSellOrder);
         List<BuyOffer> buyOffers = new ArrayList<>();
         buyOrders.forEach(buyOrder -> {
-            Optional<User> user = userDao.getUserById(buyOrder.getIdBuyer());
+            Optional<User> user = userService.getUserById(buyOrder.getIdBuyer());
             user.ifPresent(value -> buyOffers.add(new BuyOffer(buyOrder, value)));
         });
         return buyOffers;
@@ -70,8 +70,11 @@ public class BuyOrderServiceImpl implements BuyOrderService {
     @Transactional
     @Override
     public void confirmBuyOrder(int sellOrderId, int buyerId, int seller, int productId, BigDecimal price) {
+        if (!userService.currentUserOwnsNft(productId))
+            throw new UserIsNotNftOwnerException();
+
         SellOrder sOrder = sellOrderDao.getOrderById(sellOrderId).orElseThrow(SellOrderNotFoundException::new);
-        User buyerUser = userDao.getUserById(buyerId).orElseThrow(UserNotFoundException::new);
+        User buyerUser = userService.getUserById(buyerId).orElseThrow(UserNotFoundException::new);
 
         nftDao.updateOwner(sOrder.getNftId(), buyerUser.getId());
         sellOrderDao.delete(sOrder.getId());
@@ -80,6 +83,9 @@ public class BuyOrderServiceImpl implements BuyOrderService {
 
     @Override
     public void deleteBuyOrder(int sellOrderId, int buyerId) {
+        if (!userService.currentUserOwnsSellOrder(sellOrderId))
+            throw new UserIsNotNftOwnerException();
+
         buyOrderDao.deleteBuyOrder(sellOrderId, buyerId);
     }
 }
